ADDED: 3d model class

+++ LeoDeg.DXEngine/DirectXEngine/Graphics/Mesh.cpp
@@ -6,10 +6,10 @@ namespace DXEngine
 	{
 		this->m_DeviceContext = deviceContext;
 
-		HRESULT hResult = this->m_VertexBuffer.Initialize (device, vertices.data (), vertices.size ());
+		HRESULT hResult = this->m_VertexBuffer.Initialize (device, vertices.data (), static_cast<UINT>(vertices.size ()));
 		COM_ERROR_IF_FAILED (hResult, "Failed to initialize vertex buffer for mesh.");
 
-		hResult = this->m_IndexBuffer.Initilization (device, indices.data (), indices.size ());
+		hResult = this->m_IndexBuffer.Initilization (device, indices.data (), static_cast<UINT>(indices.size ()));
 		COM_ERROR_IF_FAILED (hResult, "Failed to initialize index buffer for mesh.");
 	}
